requirements.txt
# Core dependencies
numpy>=1.19.0
scipy>=1.5.0
matplotlib>=3.3.0
opencv-python>=4.5.0
tqdm>=4.60.0

# Scientific computing
scikit-learn>=0.24.0
Pillow>=8.0.0

# Data handling
pandas>=1.2.0
h5py>=3.1.0
PyYAML>=5.4.0

# Visualization
seaborn>=0.11.0
plotly>=5.0.0

# Development dependencies (optional, for dev installation)
pytest>=6.2.0
pytest-cov>=2.12.0
black>=21.0.0
flake8>=3.8.0
pre-commit>=2.13.0
mypy>=0.812

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=0.5.0

# Optional GPU acceleration
# torch>=1.9.0  # Uncomment if using PyTorch for GPU features
# torchvision>=0.10.0

# Optional ROS integration  
# rospy  # Uncomment if using ROS
# geometry_msgs
# sensor_msgs

# Optional deep learning features
# tensorflow>=2.5.0  # Uncomment if using TensorFlow features
# keras>=2.4.0